Remove implicit setting of DIR_OUTPUT when initalizing direction
In __init__() self._IODIRA = 0 and self._IODIRB = 0 causes all
pins to be set to output, implicitly.

When this first pin is set to DIR_INPUT in open(), all other
pins are DIR_OUTOUT, as self._IODIRA = 0x01.

This is not desirable.

This commit reads the state of _IODIRx, _GPIOx and _GPPUx to avoid
any implicit initializations, by fetching state from the chip.